Add scalable measurements
There are many possible units for measurements, some of which should
only be scaled in certain ways (e.g. you can't have half an egg).
Therefore, scaling a particular measurement depends on the unit of the
measurement.

Measurements are a value in a given unit. Units may have lots of
properties (such as overrides), but they are represented as strings here
so that they can easily be turned into JSON and so that that information
can be looked up and stored centrally, rather than on every unit.