Add generic function to convert between units
This uses a minimal specification of how units can be converted to
produce a function that can convert between any connected units. I.e.,
if conversions are given for A to B and B to C, the function can convert
from A to C and vice versa. The map that is built to perform this is
computationally expensive, so is the first argument to a curried
function. This map could be stored and used to preload the function to
improve performance.

This is deliberately specified in a declarative manner so that all the
specific conversion logic could be stored as JSON, making extension to
new units easy.